json file can be imported, temp display raw json.
